view_curses: do not force black background
Without a call to use_default_colors, ncurses will assume that color 0
(usually black) is the same as the background color of the terminal.

This assumption is not necessarily true, the user might be running a
terminal with any background (even white or a light color) while the
color palette has black at index 0.

To avoid overriding the background color for the entire window without a
good reason, invoke use_default_colors after start_color.